+++ src/pages/mypage/components/index.js
@@ -1,0 +1,3 @@
+export * from './purchases-list';
+export * from './my-info';
+export * from './my-point';
