+++ src/pages/mypage/components/my-info/index.js
@@ -1,0 +1,1 @@
+export * from './MyInfo';
