+++ src/pages/mypage/components/my-point/index.js
@@ -1,0 +1,1 @@
+export * from './MyPoint';
