+++ src/pages/mypage/components/purchases-list/index.js
@@ -1,0 +1,1 @@
+export * from './PurchasesList';
