+++ src/pages/mypage/mock/index.js
@@ -1,0 +1,1 @@
+export * from './purchase-data';
