+++ src/pages/mypage/ui/index.js
@@ -1,1 +1,2 @@
 export { default as Mypage } from './Mypage';
+export * from './Mypage.style';
